+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -278,7 +278,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01JY01QA00C1JP9XP9QE92MHAP"),
+  _id: z.string().default("01JY0WJ7BD1EBP1Q3570Z1ZZ5F"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,
@@ -322,7 +322,7 @@ export const ListDatasourcesData$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListDatasourcesData
 > = z.object({
-  id: z.string().default("01JY01QA00C1JP9XP9QE92MHAP"),
+  id: z.string().default("01JY0WJ7BD1EBP1Q3570Z1ZZ5F"),
   displayName: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$outboundSchema,