+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -244,7 +244,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01JY01QA01RRHH0T6M1HAJWK17"),
+  _id: z.string().default("01JY0WJ7BF3EEJ47JAVSFZ2QVT"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,
@@ -288,7 +288,7 @@ export const UpdateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01JY01QA01RRHH0T6M1HAJWK17"),
+  id: z.string().default("01JY0WJ7BF3EEJ47JAVSFZ2QVT"),
   displayName: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$outboundSchema,